requirements.txt
connexion~=1.0
requests~=2.11
pyyaml~=3.12
redis~=2.10